Add schema with fields.
